#1 Rename mis-named package v1 to json
In package names v1/v2/v${nonzerointeger} is reserved to identify module
versions. This was not the case of the previous state but rather the
version of the blueprint implementation.

This commit removes this misnaming towards json to amend the situation
above and also to put the focus on the format being worked on.